use intptr_t for int<>void* casting
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/branches/V5-3-patches@15490 06827809-a52a-0410-b366-d66718629ded

+++ net-snmp/agent/helpers/row_merge.c
@@ -31,7 +31,7 @@ netsnmp_get_row_merge_handler(int prefix_len)
     ret = netsnmp_create_handler("row_merge",
                                   netsnmp_row_merge_helper_handler);
     if (ret) {
-        ret->myvoid = (void *) prefix_len;
+        ret->myvoid = (void *)(intptr_t)prefix_len;
     }
     return ret;
 }
@@ -165,7 +165,7 @@ netsnmp_row_merge_helper_handler(netsnmp_mib_handler *handler,
      * Use the prefix length as supplied during registration, rather
      *  than trying to second-guess what the MIB implementer wanted.
      */
-    int SKIP_OID = (int)handler->myvoid;
+    int SKIP_OID = (int)(intptr_t)handler->myvoid;
 
     DEBUGMSGTL(("helper:row_merge", "Got request (%d): ", SKIP_OID));
     DEBUGMSGOID(("helper:row_merge", reginfo->rootoid, reginfo->rootoid_len));
